Rename LazzyChoice to LazyChoice
